requirements.txt
Flask==3.0.3
python-dotenv==1.0.0
google-generativeai==0.5.0
flask-sqlalchemy==3.0.5
flask-login==0.6.3
PyPDF2==3.0.1
python-docx==1.0.0
pandas==2.2.2
cryptography==42.0.5
langdetect==1.0.9
gunicorn==23.0.0
flask-wtf==1.2.2
WTForms==3.2.1
bcrypt==4.3.0
requests==2.32.5
APScheduler==3.10.4
Flask-Mail==0.9.1
APScheduler==3.10.4
cryptography==42.0.5
flask-login==0.6.3
Flask-Mail==0.9.1
flask-sqlalchemy==3.0.5
langdetect==1.0.9
PyPDF2==3.0.1
python-docx==1.0.0
werkzeug
flask-babel
